+++ src/FingersCrossedProcessor.php
@@ -1,19 +1,14 @@
 <?php
+declare(strict_types=1);
 
 namespace League\Pipeline;
 
 class FingersCrossedProcessor implements ProcessorInterface
 {
-    /**
-     * @param array $stages
-     * @param mixed $payload
-     *
-     * @return mixed
-     */
-    public function process(array $stages, $payload)
+    public function process($payload, callable ...$stages)
     {
         foreach ($stages as $stage) {
-            $payload = call_user_func($stage, $payload);
+            $payload = $stage($payload);
         }
 
         return $payload;