+++ src/ProcessorInterface.php
@@ -1,14 +1,16 @@
 <?php
+declare(strict_types=1);
 
 namespace League\Pipeline;
 
 interface ProcessorInterface
 {
     /**
-     * @param array $stages
+     * Process the payload using multiple stages.
+     *
      * @param mixed $payload
      *
      * @return mixed
      */
-    public function process(array $stages, $payload);
+    public function process($payload, callable ...$stages);
 }